requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
google-api-python-client==2.110.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
streamlit==1.28.2
plotly==5.17.0
pandas==2.1.4